requirements.txt
discord.py==2.6.3
aiosqlite==0.21.0
aiohttp==3.12.15
APScheduler==3.11.0
python-dotenv==1.1.1
pytest==8.4.2
pytest-asyncio==1.2.0
pynacl==1.6.0